refactor: 역할에 맞게 폴더 분류
